Use LLVM Flang toolchain only for CPU builds
Cuda 9.2 misses LLVM 7 support. Switch cuda builds in a separate commit, to
simplify backporting of CPU changes.

+++ tools/pip/setup.py
@@ -153,8 +153,21 @@ if variant != 'NATIVE':
                     os.path.join(CURRENT_DIR, 'mxnet/include/mkldnn'))
 if platform.system() == 'Linux':
     libdir, mxdir = os.path.dirname(LIB_PATH[0]), os.path.join(CURRENT_DIR, 'mxnet')
-    shutil.copy(os.path.join(libdir, 'libomp.so'), mxdir)
-    package_data['mxnet'].append('mxnet/libomp.so')
+    if os.path.exists(os.path.join(libdir, 'libgfortran.so.3')):
+        shutil.copy(os.path.join(libdir, 'libgfortran.so.3'), mxdir)
+        package_data['mxnet'].append('mxnet/libgfortran.so.3')
+    else:
+        shutil.copy(os.path.join(libdir, 'libgfortran.so.4'), mxdir)
+        package_data['mxnet'].append('mxnet/libgfortran.so.4')
+    if os.path.exists(os.path.join(libdir, 'libquadmath.so.0')):
+        shutil.copy(os.path.join(libdir, 'libquadmath.so.0'), mxdir)
+        package_data['mxnet'].append('mxnet/libquadmath.so.0')
+    if os.path.exists(os.path.join(libdir, 'libopenblas.so.0')):
+        shutil.copy(os.path.join(libdir, 'libopenblas.so.0'), mxdir)
+        package_data['mxnet'].append('mxnet/libquadmath.so.0')
+    if os.path.exists(os.path.join(libdir, 'libomp.so')):
+        shutil.copy(os.path.join(libdir, 'libomp.so'), mxdir)
+        package_data['mxnet'].append('mxnet/libomp.so')
 
 # Copy licenses and notice
 for f in os.listdir('mxnet/licenses'):
